Module 2 changes
added list users and post users methods, which are used on the reqres.in api. It uses RESTful api standards. Doesn't store data, so it is only really used for testing.

Some comments added to aid in explanation to future self when reading back old projects.

Other comments to show what was tried in the past that did not end up working (i.e. computed fetch data, as it only ended up working when put into a method).