Folder and file prep
preparing to add a header section where a search bar would be for searching for certain content in the app. Not sure how I would implement it just yet, but the concept and idea is sort of there for how the site should look like in the end.

Also perhaps for the end content, so the main content components of each page, perhaps they should mainly be in grid layout, and if I really want to, perhaps flexbox because these 2 layouts are really powerful and cool sounding to use.