updating the vue components
updating the vue components to fit better with the bigger screen of the upstairs monitor and also use some more relative elements. also made the content frame have a transition when it changed in size so it didn't look so weird when it was changing size.

also there is the things with the axios and vue.prototype, looks like it doesn't want to work for me with either vite nor vue cli, so that is something that I should not hold against vite anymore